+++ backported/COPYRIGHT
